requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
argon2-cffi==21.3.0
asgiref==3.5.2; python_version >= '3.7'
cffi==1.15.1
cycler==0.11.0; python_version >= '3.6'
django-cors-headers==3.13.0
django==4.1
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
faker==13.15.1
fonttools==4.34.4; python_version >= '3.7'
kiwisolver==1.4.4; python_version >= '3.7'
matplotlib==3.5.3
numpy==1.23.1; python_version >= '3.8'
packaging==21.3; python_version >= '3.6'
pillow==9.2.0
pycparser==2.21
pyjwt==2.4.0; python_version >= '3.6'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.2
randimage==1.2
scipy==1.9.0; python_version < '3.12' and python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tqdm==4.64.0
